refactor: cleanup camera view
